refactor(transform): split file-tree feature wiring tests
Keep features.rs under the 350-line new-file limit after adding file-tree preprocess hooks.
